refactor(github-issues-runtime): split demo-index helpers and tests (#1387)
